pyproject.toml
[project]
name = "autopilot"
version = "0.1.0"
description = "Model-agnostic CLI tool that runs AI automations on a schedule"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "anyio>=4.0.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.0",
  "rich>=13.7.0",
  "typer>=0.12.3",
]

[project.optional-dependencies]
claude-sdk = ["claude-agent-sdk>=0.1.44"]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.12.7",
]

[project.scripts]
autopilot = "autopilot.cli:app"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "W"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/autopilot"]
